+++ src/cobalt_model/mod.rs
@@ -6,7 +6,6 @@ mod mark;
 mod sass;
 mod site;
 mod template;
-mod vwiki;
 
 pub mod files;
 pub mod pagination;
@@ -33,5 +32,3 @@ pub use self::sass::SassCompiler;
 pub use self::site::Site;
 pub use self::template::Liquid;
 pub use self::template::LiquidBuilder;
-pub use self::vwiki::Vimwiki;
-pub use self::vwiki::VimwikiBuilder;