+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.19.0] - 2023-06-06
+
+### Compatibility
+
+- Removed support for vimwiki (parser/renderer was unmaintained)
+- MSRV is now 1.70.0
+
+## [0.18.5] - 2023-06-02
+
+### Fixes
+
+- Include frontmatter tags in jsonfeed
+
 ## [0.18.4] - 2023-04-20
 
 ### Fixes
@@ -670,7 +683,9 @@ this.
 [@tak1n]: https://github.com/tak1n
 
 <!-- next-url -->
-[Unreleased]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.4...HEAD
+[Unreleased]: https://github.com/cobalt-org/cobalt.rs/compare/v0.19.0...HEAD
+[0.19.0]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.5...v0.19.0
+[0.18.5]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.4...v0.18.5
 [0.18.4]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.3...v0.18.4
 [0.18.3]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.2...v0.18.3
 [0.18.2]: https://github.com/cobalt-org/cobalt.rs/compare/v0.18.1...v0.18.2