+++ crates/file-serve/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.3.0] - 2023-08-22
+
+### Breaking Change
+
+- `serve` now returns `()`
+
+### Features
+
+- `serve` can now be canceled
+
+## [0.2.4] - 2023-06-06
+
+### Compatibility
+
+- MSRV is now 1.70.0
+
+## [0.2.3] - 2023-06-02
+
 ## [0.2.2] - 2023-04-20
 
 ### Compatibility
@@ -24,7 +42,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.1] - 2022-04-04
 
 <!-- next-url -->
-[Unreleased]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.2...HEAD
+[Unreleased]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.3.0...HEAD
+[0.3.0]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.4...file-serve-v0.3.0
+[0.2.4]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.3...file-serve-v0.2.4
+[0.2.3]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.2...file-serve-v0.2.3
 [0.2.2]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.1...file-serve-v0.2.2
 [0.2.1]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.2.0...file-serve-v0.2.1
 [0.2.0]: https://github.com/cobalt-org/cobalt.rs/compare/file-serve-v0.1.1...file-serve-v0.2.0